requirements.txt
flask==3.0.3
pymongo==4.7.3
python-dotenv==1.0.1
dnspython==2.6.1
certifi==2024.2.2
selenium==4.21.0
pytest==8.2.2
webdriver-manager
webdriver-manager
prometheus-flask-exporter
